Reorganize textinput interface
