+++ app/core/UIManager.js
@@ -11,7 +11,7 @@ define(function(require, exports, module) {
     require('core/interfaces/_internals/permissions/interface'),
     require('core/interfaces/_internals/views/interface'),
     require('core/interfaces/_internals/file_preview/interface'),
-    require('core/interfaces/textinput/index'),
+    require('core/interfaces/textinput/component'),
     require('core/interfaces/section_break/section_break'),
     require('core/interfaces/checkbox/checkbox'),
     require('core/interfaces/color/index'),
